Fix an error with the Flameshot configuration
A lot of the configurations in /etc/skel were copied over from my personal Arch Linux machine. Trying to use Flameshot in Meowix, as a result, would throw an error, as the screenshot directory was set to the absolute path to my home directory.
Thus, I have removed the screenshot directory line from the Flameshot configuration. Unfortunately, it seems you cannot use ~ or $HOME in the Flameshot configuration.